For repeatable actions, make start of next cycle configurable
Previously the starting point of the next cycle was hard-coded: If you
used as action one of mark, outline-mark-subtree, or org-mark-subtree,
then the cycle for the next action started at region; otherwise it
started at the same type as the current target. There was a TODO in
the source suggesting an extra customization variable, but since this
option only makes sense for repeatable actions, I decided to
generalize the items in embark-repeat-actions to allow pairs of the
form (action . next-type).