Have embark-{next,previous}-symbol act on symbol at point
The old way, which prompted for a symbol, was very similar to existing
commands like isearch-forward-symbol or word-search-forward; and in
particular I couldn't really imagine anyone wanting to use the embark
versions directly (i.e., not through embark-act) given the existence
of isearch-forward-symbol, which even has a default binding. But these
new versions that act on the symbol at point are something you might
actually want to bind directly. And for the usage as actions they work
exactly as before.