Improve handling of relative paths in export, collect and act-all
Before this change I had some ad hoc handling of converting from
relative to absolute paths (see the two removed TODOs) that failed to
cover some cases where it was needed. For example, I trie attaching
multiple files to an email message by using embark-act-all from
mml-attach-file, and this failed to find the files if they weren't in
the default-directory of the email composition buffer!

The basic source of the problem is that typically file completion
candidates are just the last portion of a file path, and should be
taken relative to the directory specified in the minibuffer input, but
that directory is *not* necessarily the default-directory of the
minibuffer or the buffer current before the minibuffer. Embark has
long had embark--default-directory which has some reasonable
heuristics to figure out the correct directory candidates which are
relative paths can be found in.

With this change all use of candidate collectors (those are used for
all the commands for sets of candidates: export, collect and act-all)
will now add the correct path to file candidates, so that future
actions can safely find them.

Previously dired export would make candidates relative to an
appropriately chosen directory to achieve a less cluttered look, but
collect buffers did not do this. Now collect buffers do so as well.