Incomplete implementation of revert in export buffers
This partially addresses #171.

It has several problems still:

1. It doesn't work for consult-async commands since (1) the target
buffer is usually some buffer shown by consult preview that is dead
by then, (2) in inserts the entire minibuffer contents so you wind up
with double #, (3) it doesn't wait for the candidate generation to
finish.

2. Since the exporters create *and* display their own buffers, and for
some this might be hard to avoid, I currently kill the exported buffer
and generate it anew, which usually displays it somewhere else. :(

3. Some export buffers bind g to something other than revert-buffer so
the behavior is surprising, for example, grep export buffers seems to
have g bound to recompile.

4. It should probably be generalized to cover embark-collect buffers
too.

So this is definitely just a proof of concept. I'm pushing it since it
just adds invisible functionality, even if that functionality isn't
what we want in all cases. And hey, it does work for embark export
from a non-mini buffer, for file export (which is great to update
after renaming or deleting files!) and for occur export from
consult-line.