Implement "revert = just rerun" for export
This removes the previous revert behavior with its hacky "peek at
Consult's minibuffer overlay to figure out when the search is over" in
favor of simply killing the export buffer and rerunning the command.

This simpler behavior was suggested by @minad in #171.

Still left to do: implement it for collect as well.