Simplify collect candidate storage, stop using revert
Remove the variable embark-collect--candidates. The entries were
duplicated in variable tabulated-list-entries anyway, now we only
store them there.

Also, instead of storing the original candidate (with tofu and
whatnot) in a text property, we now store it as the id of the
tabulated list entry.

Finally, some internal things used to rely on revert-buffer being
tabulated-list-revert. We now call tabulated-list-revert or
tabulated-list-print directly as appropriate. This will allow use to
change the value of revert-buffer-function to mean "rerun the command
that created this collect buffer".