Allow functions of no arguments as action hooks
It always bothered me that we had some trivial wrappers just because
action hooks were forced to take :action, :target, :type parameters.
With this change we can eliminate three wrappers:

- embark--xref-push-marker for xref-push-marker-stack
- embark--unmark-target for deactivate-mark
- embark--force-complete for minibuffer-force-complete

As always in Embark, I like increasing the odds that a pre-existing
function can be used in Embark without wrapping.