Remove embark-eshell, add more general embark--cd
The new embark--cd pre-action hook will cause the action to run with
default-directory set to the directory associated to the target, a
file, buffer, bookmark or library.

Having that allows us to remove embark-eshell, instead using eshell
directly. It also makes sense know to bind eshell in the buffer and
bookmark maps as well. In case people prefer shell over eshell, we
have embark--cd setup it for it too.