Add "obey/disobey embark-quit-after-action" message
Whether or not an action will quit the minibuffer is the xor of three
boolean values:

- Whether there was a prefix argument.
- Whether the user pressed q an odd number of times.
- What (alist-get action embark-quit-after-action) says.

We cannot know the last one until the user selects the action, of
course, but the other two are known before then. So after the prefix
argument and after each press of q we can tell the user whether we
will obey or disobey embark-quit-after-action.