Always transform candidates for multitarget actions
There are 3 ways you can call the default action: via embark-dwim,
embark-act, or embark-act-all, and they should all behave the same
way. In particular, for a multitarget action, they should all behave
like embark-act-all. But embark-act-all was using transformed
candidates while embark-act and embark-dwim were using original
candidates.

The use of original candidates is meant for actions called
interactively, and doesn't really make sense for multitarget actions
which are not called interactively. So in the case of multitarget
actions it was embark-act-all who was right and this change makes
embark-act and embark-dwim agree with it.

This should fix the problem over at bdarcus/citar#583.