Implement a general Org target finder
This replaces the link and table-cell target finders.

The functionality of the old link target finder is now implemented via
a transformer.

I could let it recognize EVERYTHING, but I think it's probably wisest
to only recognize types for which we have actions, and maybe a few
cases of types with no actions but for which we want w, DEL and i.