Fix #515
In terminal Emacs, the default binding for scroll-other-window, namely
C-M-v, is read as ESC C-v, i.e., [27 22]. When you lookup this in a
keymap where it is not defined using lookup-key, you do not get nil,
but rather the number 1, indicating that [27] is an invalid prefix. So
I really should have been checking for a non-nil, *non-number* return
value to indicate the key is directly bound.