Don't rely on func-arity since advice can change it
It's not rare to have :around advice with parameter list of the
form:(fn &rest args) which would change the return value of func-arity
to (0 . many), so func-arity is not a reliable way to find out the
number of arguments eww-download-callback takes. Thanks to Stefan
Monnier for pointing this out.

I did a quick check of Emacs versions and 27.2 still had the
two-argument version of the function, whereas 28.0.90 has the
three-argument version. So, following Stefan's suggestion, I'll check
for emacs-major-version ≥ 28.