BROKEN -- Add embark--unfold -- BROKEN
Do not merge!

Add an embark--unfold post action to open closed overlays. Unfortunately it
doesn't work yet as expected in Org files. I don't know if this is the fault of
Embark or of the new Org fold API. Unfortunately the new Org fold API is
problematic. While the idea isn't bad, I only had issues with it.

Fix #571