Run all actions with the default-directory of the originating buffer
There are two buffers relevant for embark-act: the buffer in which
you run embark-act (here called the "originating buffer") and the
target buffer, the buffer in which embark-act will run the action.

These two buffers agree for actions from "regular" buffers, where
"regular" means not the minibuffer and not an Embark Collect buffer.
When you call embark-act in the minibuffer, the action is run in the
buffer that was current when you ran the command that opened the
minibuffer, and when you call embark-act from an Embark Collect
buffer, the target buffer is the buffer that was current when you ran
the command that produced the candidates in the first place.

OK, enough background: with this change, when running an action in the
target buffer, default-directory will be let-bound to the value of
default-directory in the originating buffer. This fixes #612 and, I
hope, doesn't mess anything else up.