Clip org targets to narrowed portion of buffer
The bounds returned by org-element-context give the full org element,
ignoring any narrowing. This caused an error getting the target string
with buffer-substring.

I now clip the org element bounds to the narrowed portion of the
buffer. For actions that are org commands, this will make no
difference since they do whatever they do, we don't pass them the
embark bounds or the embark target string. For actions ike saving on
the kill-ring, it'll be less confusing if the target and bounds
correspond to what is visible in the narrowed buffer.

This fixes #614.