Make `embark-export` abide to `display-buffer-alist`
This change makes it so that `buffer` is named before it is
`pop-to-buffer`'d to, so that the rules specified `display-buffer-alist`
can be matched by the Embark export buffer name patterns.

E.g., the following addition makes Embark export buffers show in a
side window.
```elisp
(add-to-list 'display-buffer-alist '("\\*Embark Export: .*"
                                     (display-buffer-in-side-window)
```