Allow candidate collectors to include bounds
Previously candidate collectors had to return something of the form
(cons TYPE STRINGS):, now they can return (cons TYPE LIST) where each
element of LIST is either a STRING or of the form (cons STRING
BOUNDS).

These bounds are used by commands that act on collected candidates as
follows:

- embark-collect and embark-export ignore the bounds.
- embark-act will pass the bounds to the action's hooks.

The following candidate collectors have been updated to include bounds
information:

- embark-selected-candidates
- embark-completions-buffer-candidates
- embark-dired-candidates
- embark-embark-collect-candidates
- embark-custom-candidates

Internal changes:

Previously embark-selected-candidates already passed bounds
information, but did so in an indirect fashion, attaching an overlay
as a text property to the candidates. With the new functionality this
has been simplified: now embark--selection is an alist associating
selected candidates to overlays (or nil, if the candidate has no
bounds), but actions do not recieve the overlays, only the bounds (as
markers, not numbers; and the markers are carefully GC-ed).