Make org block actions work with selections
For actions that work on the thing at point, you need to move to each
of the target for the action to work properly with selections. Maybe
this requires a more global intervention, but for now I'm fixing the
actions I notice that need it. (For example, actions that really work
on the region are OK, because they already used embark--mark-target).