Free up i for embark-insert in more cases
I feel a little bad about changing the binding for indent-rigidly,
because it is well-established now, but since that command uses the
arrow keys anyway, binding it to left and right just makes sense.

The point for me is that once you get used to duplicating things at
point with C-. i, it gets really frustrating when C-. i does something
else instead!