Run actions in minibuffer if not in a completion session
If you are in a minibuffer completion session, embark runs actions in
the previously selected buffer. Unfortunately, embark was also doing
that if you are editing a minibuffer that is not in a completion
session, making it impossible to use embark for normal editing tasks in
the minibuffer (such as deleting an s-expression, replacing an
s-expression with its value, etc.).