Give explicit file paths higher priority than ffap-guessed
With this change when a bare identifier is guessed by ffap to
represent a file, you get the identifier target first, and later the
file. This fixes #673.

Previously the file target finder was also split in two parts, with
the purpose of making elisp library files have lower priority than
elisp symbols. With this change the high priority file finder is
tightened up to only match file names with a path separator or with an
extension, and the lower priority elisp library file target finder is
now generalized to match any indentifier that ffap thinks is a file.