Use org-element only in org-mode buffers (fixes #698)
The org-element parser will start raising errors in non-org-mode
buffers in future org releases. We were using it in org-agenda buffers
too, but the benefit is almost non-existent: by far the most
important org targets in agenda buffers are headings and links and
those target finders do not use org-element.