fix: limit set--this-command-keys to browse-url
My view after investigating this #470 and #182 is that the underlying
cause is a bug in `browse-url-interactive-arg` itself, as it should
handle `this-command-keys` being nil. However, even if that is fixed,
older versions of Emacs will still have the issue. So I propose limiting
the use of `set--this-command-keys` to the functions with `browse-url`
in their name.